Added Chat

+++ lib/util/size_config.dart
@@ -19,8 +19,7 @@ class SizeConfig {
 double getProportionateScreenHeight(double inputHeight) {
   double screenHeight = SizeConfig.screenHeight;
   // 812 is the layout height that designer use
-  //convert result to double
-  return ((inputHeight / 812.0) * screenHeight).toDouble();
+  return (inputHeight / 812.0) * screenHeight;
 }
 
 // Get the proportionate height as per screen size
